foodrescue-agent/requirements.txt
streamlit==1.39.0
google-generativeai==0.8.0
python-dotenv==1.0.1
plotly==6.1.1
kaleido==1.2.0
pandas==2.0.3
requests==2.32.3
jsonschema==4.21.1
typing-extensions==4.11.0
python-dateutil==2.9.0.post0
